+++ packages/core/src/index.ts
@@ -13,18 +13,9 @@ const DISPOSED = 1 << 3;
 const HAS_ERROR = 1 << 4;
 const TRACKING = 1 << 5;
 
-// Flags for Nodes.
-const NODE_FREE = 1 << 0;
-const NODE_SUBSCRIBED = 1 << 1;
-
 // A linked list node used to track dependencies (sources) and dependents (targets).
 // Also used to remember the source's last version number that the target saw.
 type Node = {
-	// A node may have the following flags:
-	//  NODE_FREE when it's unclear whether the source is still a dependency of the target
-	//  NODE_SUBSCRIBED when the target has subscribed to listen change notifications from the source
-	_flags: number;
-
 	// A source whose value the target depends on.
 	_source: Signal;
 	_prevSource?: Node;
@@ -38,6 +29,7 @@ type Node = {
 	// The version number of the source that target has last seen. We use version numbers
 	// instead of storing the source value, because source values can take arbitrary amount
 	// of memory, and computeds could hang on to them forever because they're lazily evaluated.
+	// Use the special value -1 to mark potentially unused but recyclable nodes.
 	_version: number;
 
 	// Used to remember & roll back the source's previous `._node` value when entering &
@@ -121,19 +113,32 @@ function addDependency(signal: Signal): Node | undefined {
 
 	let node = signal._node;
 	if (node === undefined || node._target !== evalContext) {
-		// `signal` is a new dependency. Create a new node dependency node, move it
-		//  to the front of the current context's dependency list.
+		/**
+		 * `signal` is a new dependency. Create a new dependency node, and set it
+		 * as the tail of the current context's dependency list. e.g:
+		 *
+		 * { A <-> B       }
+		 *         ↑     ↑
+		 *        tail  node (new)
+		 *               ↓
+		 * { A <-> B <-> C }
+		 *               ↑
+		 *              tail (evalContext._sources)
+		 */
 		node = {
-			_flags: 0,
 			_version: 0,
 			_source: signal,
-			_prevSource: undefined,
-			_nextSource: evalContext._sources,
+			_prevSource: evalContext._sources,
+			_nextSource: undefined,
 			_target: evalContext,
 			_prevTarget: undefined,
 			_nextTarget: undefined,
 			_rollbackNode: node,
 		};
+
+		if (evalContext._sources !== undefined) {
+			evalContext._sources._nextSource = node;
+		}
 		evalContext._sources = node;
 		signal._node = node;
 
@@ -143,26 +148,34 @@ function addDependency(signal: Signal): Node | undefined {
 			signal._subscribe(node);
 		}
 		return node;
-	} else if (node._flags & NODE_FREE) {
-		// `signal` is an existing dependency from a previous evaluation. Reuse the dependency
-		// node and move it to the front of the evaluation context's dependency list.
-		node._flags &= ~NODE_FREE;
+	} else if (node._version === -1) {
+		// `signal` is an existing dependency from a previous evaluation. Reuse it.
+		node._version = 0;
 
-		const head = evalContext._sources;
-		if (node !== head) {
-			const prev = node._prevSource;
-			const next = node._nextSource;
-			if (prev !== undefined) {
-				prev._nextSource = next;
+		/**
+		 * If `node` is not already the current tail of the dependency list (i.e.
+		 * there is a next node in the list), then make the `node` the new tail. e.g:
+		 *
+		 * { A <-> B <-> C <-> D }
+		 *         ↑           ↑
+		 *        node   ┌─── tail (evalContext._sources)
+		 *         └─────│─────┐
+		 *               ↓     ↓
+		 * { A <-> C <-> D <-> B }
+		 *                     ↑
+		 *                    tail (evalContext._sources)
+		 */
+		if (node._nextSource !== undefined) {
+			node._nextSource._prevSource = node._prevSource;
+
+			if (node._prevSource !== undefined) {
+				node._prevSource._nextSource = node._nextSource;
 			}
-			if (next !== undefined) {
-				next._prevSource = prev;
-			}
-			if (head !== undefined) {
-				head._prevSource = node;
-			}
-			node._prevSource = undefined;
-			node._nextSource = head;
+
+			node._prevSource = evalContext._sources;
+			node._nextSource = undefined;
+
+			evalContext._sources!._nextSource = node;
 			evalContext._sources = node;
 		}
 
@@ -177,7 +190,11 @@ declare class Signal<T = any> {
 	/** @internal */
 	_value: unknown;
 
-	/** @internal */
+	/**
+	 * @internal
+	 * Version numbers should always be >= 0, because the special value -1 is used
+	 * by Nodes to signify potentially unused but recyclable nodes.
+	 */
 	_version: number;
 
 	/** @internal */
@@ -222,10 +239,8 @@ Signal.prototype._refresh = function () {
 };
 
 Signal.prototype._subscribe = function (node) {
-	if (!(node._flags & NODE_SUBSCRIBED)) {
-		node._flags |= NODE_SUBSCRIBED;
+	if (this._targets !== node && node._prevTarget === undefined) {
 		node._nextTarget = this._targets;
-
 		if (this._targets !== undefined) {
 			this._targets._prevTarget = node;
 		}
@@ -234,9 +249,8 @@ Signal.prototype._subscribe = function (node) {
 };
 
 Signal.prototype._unsubscribe = function (node) {
-	if (node._flags & NODE_SUBSCRIBED) {
-		node._flags &= ~NODE_SUBSCRIBED;
-
+	// Only run the unsubscribe step if the signal has any subscribers to begin with.
+	if (this._targets !== undefined) {
 		const prev = node._prevTarget;
 		const next = node._nextTarget;
 		if (prev !== undefined) {
@@ -325,8 +339,11 @@ function needsToRecompute(target: Computed | Effect): boolean {
 	// Check the dependencies for changed values. The dependency list is already
 	// in order of use. Therefore if multiple dependencies have changed values, only
 	// the first used dependency is re-evaluated at this point.
-	let node = target._sources;
-	while (node !== undefined) {
+	for (
+		let node = target._sources;
+		node !== undefined;
+		node = node._nextSource
+	) {
 		// If there's a new version of the dependency before or after refreshing,
 		// or the dependency has something blocking it from refreshing at all (e.g. a
 		// dependency cycle), then we need to recompute.
@@ -335,16 +352,27 @@ function needsToRecompute(target: Computed | Effect): boolean {
 			!node._source._refresh() ||
 			node._source._version !== node._version
 		) {
-			break;
+			return true;
 		}
-		node = node._nextSource;
 	}
-	// If none of the dependencies have changed values since last recompute then the
+	// If none of the dependencies have changed values since last recompute then
 	// there's no need to recompute.
-	return node !== undefined;
+	return false;
 }
 
 function prepareSources(target: Computed | Effect) {
+	/**
+	 * 1. Mark all current sources as re-usable nodes (version: -1)
+	 * 2. Set a rollback node if the current node is being used in a different context
+	 * 3. Point 'target._sources' to the tail of the doubly-linked list, e.g:
+	 *
+	 *    { undefined <- A <-> B <-> C -> undefined }
+	 *                   ↑           ↑
+	 *                   │           └──────┐
+	 * target._sources = A; (node is head)  │
+	 *                   ↓                  │
+	 * target._sources = C; (node is tail) ─┘
+	 */
 	for (
 		let node = target._sources;
 		node !== undefined;
@@ -355,40 +383,67 @@ function prepareSources(target: Computed | Effect) {
 			node._rollbackNode = rollbackNode;
 		}
 		node._source._node = node;
-		node._flags |= NODE_FREE;
+		node._version = -1;
+
+		if (node._nextSource === undefined) {
+			target._sources = node;
+			break;
+		}
 	}
 }
 
 function cleanupSources(target: Computed | Effect) {
-	// At this point target._sources is a mishmash of current & former dependencies.
-	// The current dependencies are also in a reverse order of use.
-	// Therefore build a new, reverted list of dependencies containing only the current
-	// dependencies in a proper order of use.
-	// Drop former dependencies from the list and unsubscribe from their change notifications.
-
 	let node = target._sources;
-	let sources = undefined;
+	let head = undefined;
+
+	/**
+	 * At this point 'target._sources' points to the tail of the doubly-linked list.
+	 * It contains all existing sources + new sources in order of use.
+	 * Iterate backwards until we find the head node while dropping old dependencies.
+	 */
 	while (node !== undefined) {
-		const next = node._nextSource;
-		if (node._flags & NODE_FREE) {
+		const prev = node._prevSource;
+
+		/**
+		 * The node was not re-used, unsubscribe from its change notifications and remove itself
+		 * from the doubly-linked list. e.g:
+		 *
+		 * { A <-> B <-> C }
+		 *         ↓
+		 *    { A <-> C }
+		 */
+		if (node._version === -1) {
 			node._source._unsubscribe(node);
-			node._nextSource = undefined;
-		} else {
-			if (sources !== undefined) {
-				sources._prevSource = node;
+
+			if (prev !== undefined) {
+				prev._nextSource = node._nextSource;
 			}
-			node._prevSource = undefined;
-			node._nextSource = sources;
-			sources = node;
+			if (node._nextSource !== undefined) {
+				node._nextSource._prevSource = prev;
+			}
+		} else {
+			/**
+			 * The new head is the last node seen which wasn't removed/unsubscribed
+			 * from the doubly-linked list. e.g:
+			 *
+			 * { A <-> B <-> C }
+			 *   ↑     ↑     ↑
+			 *   │     │     └ head = node
+			 *   │     └ head = node
+			 *   └ head = node
+			 */
+			head = node;
 		}
 
 		node._source._node = node._rollbackNode;
 		if (node._rollbackNode !== undefined) {
 			node._rollbackNode = undefined;
 		}
-		node = next;
+
+		node = prev;
 	}
-	target._sources = sources;
+
+	target._sources = head;
 }
 
 declare class Computed<T = any> extends Signal<T> {
@@ -435,7 +490,7 @@ Computed.prototype._refresh = function () {
 	this._globalVersion = globalVersion;
 
 	// Mark this computed signal running before checking the dependencies for value
-	// changes, so that the RUNNIN flag can be used to notice cyclical dependencies.
+	// changes, so that the RUNNING flag can be used to notice cyclical dependencies.
 	this._flags |= RUNNING;
 	if (this._version > 0 && !needsToRecompute(this)) {
 		this._flags &= ~RUNNING;
@@ -485,18 +540,22 @@ Computed.prototype._subscribe = function (node) {
 };
 
 Computed.prototype._unsubscribe = function (node) {
-	Signal.prototype._unsubscribe.call(this, node);
+	// Only run the unsubscribe step if the computed signal has any subscribers.
+	if (this._targets !== undefined) {
+		Signal.prototype._unsubscribe.call(this, node);
 
-	// Computed signal unsubscribes from its dependencies from it loses its last subscriber.
-	if (this._targets === undefined) {
-		this._flags &= ~TRACKING;
+		// Computed signal unsubscribes from its dependencies when it loses its last subscriber.
+		// This makes it possible for unreferences subgraphs of computed signals to get garbage collected.
+		if (this._targets === undefined) {
+			this._flags &= ~TRACKING;
 
-		for (
-			let node = this._sources;
-			node !== undefined;
-			node = node._nextSource
-		) {
-			node._source._unsubscribe(node);
+			for (
+				let node = this._sources;
+				node !== undefined;
+				node = node._nextSource
+			) {
+				node._source._unsubscribe(node);
+			}
 		}
 	}
 };
@@ -603,13 +662,13 @@ function endEffect(this: Effect, prevContext?: Computed | Effect) {
 }
 
 declare class Effect {
-	_compute?: () => unknown;
-	_cleanup?: unknown;
+	_compute?: () => void | (() => void);
+	_cleanup?: () => void;
 	_sources?: Node;
 	_nextBatchedEffect?: Effect;
 	_flags: number;
 
-	constructor(compute: () => void);
+	constructor(compute: () => void | (() => void));
 
 	_callback(): void;
 	_start(): () => void;
@@ -617,7 +676,7 @@ declare class Effect {
 	_dispose(): void;
 }
 
-function Effect(this: Effect, compute: () => void) {
+function Effect(this: Effect, compute: () => void | (() => void)) {
 	this._compute = compute;
 	this._cleanup = undefined;
 	this._sources = undefined;
@@ -628,8 +687,12 @@ function Effect(this: Effect, compute: () => void) {
 Effect.prototype._callback = function () {
 	const finish = this._start();
 	try {
-		if (!(this._flags & DISPOSED) && this._compute !== undefined) {
-			this._cleanup = this._compute();
+		if (this._flags & DISPOSED) return;
+		if (this._compute === undefined) return;
+
+		const cleanup = this._compute();
+		if (typeof cleanup === "function") {
+			this._cleanup = cleanup;
 		}
 	} finally {
 		finish();
@@ -667,9 +730,14 @@ Effect.prototype._dispose = function () {
 	}
 };
 
-function effect(compute: () => unknown): () => void {
+function effect(compute: () => void | (() => void)): () => void {
 	const effect = new Effect(compute);
-	effect._callback();
+	try {
+		effect._callback();
+	} catch (err) {
+		effect._dispose();
+		throw err;
+	}
 	// Return a bound function instead of a wrapper like `() => effect._dispose()`,
 	// because bound functions seem to be just as fast and take up a lot less memory.
 	return effect._dispose.bind(effect);